+++ medications_reminder_app/lib/ui/splash_screen.dart
@@ -1,4 +1,8 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
+import 'package:medications_reminder_app/navigation/app_navigation/navigation.dart';
+import 'package:medications_reminder_app/ui/home_screen.dart';
 
 
 //Note that the colors are #40b26d(main colour) and sub colour #fdfcff
@@ -6,16 +10,73 @@ import 'package:flutter/material.dart';
 //I already added the google fonts package, use poppins
 //I'M COUNTING ON YOU!!!
 class SplashScreen extends StatefulWidget {
+  // This widget is the root of your application.
   @override
   _SplashScreenState createState() => _SplashScreenState();
 }
 
 class _SplashScreenState extends State<SplashScreen> {
   @override
+  void initState() {
+    super.initState();
+    Timer(Duration(seconds: 5),() => Navigation().pushTo(context, HomeScreen()));
+  }
+  @override
   Widget build(BuildContext context) {
-    return Scaffold(
-          body: Center(
-          child: Text('Splash Screen')
+    return new Scaffold(
+      body: Stack(
+        fit: StackFit.expand,
+        children: <Widget>[
+          Container(
+            decoration: BoxDecoration(color: Colors.green[900]),
+            ),
+            Column(
+              mainAxisAlignment: MainAxisAlignment.start,
+              children: <Widget>[
+                Expanded(
+                  flex: 2,
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      CircleAvatar(
+                        backgroundColor: Colors.white,
+                        radius: 50.0,
+                        child: Icon(
+                          Icons.local_pharmacy,
+                          color: Colors.green[900],
+                          size: 50.0,
+                        ),
+                        ),
+                        Padding(
+                          padding: EdgeInsets.only(top: 10.0),
+                        ),
+                        Text(
+                          "MedBuzz",
+                        style: TextStyle(
+                          color:Colors.white,
+                          fontSize: 24.0,
+                          fontWeight: FontWeight.bold),
+                          )
+                    ],
+                  ),
+                  ),
+                  Expanded(
+                    flex: 1,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        CircularProgressIndicator(),
+                        Padding(
+                          padding: EdgeInsets.only(top: 20.0),
+                          ),
+                          Text("Reminder You \n    Can Trust", style: TextStyle(color: Colors.white, fontSize: 18.0, fontWeight: FontWeight.bold),
+                          )
+                      ]
+                    ),
+                    )
+              ]
+            )
+        ]
       ),
     );
   }
